pyproject.toml
[tool.poetry]
name = "eo-models"
version = "0.1.0"
description = ""
authors = ["Sunny Joshi <sunny.joshi@ed.ac.uk>"]
readme = "README.md"
packages = [{include = "eo_lib"}]

[tool.poetry.dependencies]
python = "3.11.*"
einops = "0.7.0"
setuptools = "69.0.0"
fiona = "1.9.5"
kornia = "0.7.0"
matplotlib = "3.8.2"
numpy = "1.26.2"
Pillow = "10.1.0"
pyproj = "3.6.1"
rasterio = "1.3.9"
Rtree = "1.1.0"
shapely = "2.0"
timm = "0.9.2"
fastapi = "^0.75"
gunicorn = "^20.1.0"
loguru = "^0.6.0"
lightning = "^2.1.2"
rapids = "^0.0.1"
torchmetrics = "^1.2.1"
pynvml = "^11.5.0"
numba = "^0.58.1"
polars = "^0.19.19"
torchgeo = "^0.5.1"
transformers = "^4.36.1"
wandb = "^0.16.1"
tensorboard = "^2.15.1"
scikit-learn = "^1.3.2"
yacs = "^0.1.8"
opencv-python = "^4.9.0.80"
deepspeed = "^0.13.2"
h5py = "^3.11.0"
optuna = "^3.6.1"
keyring = "21.3.0"
pytorch-lightning-bolts = "^0.3.2.post1"
ray = {extras = ["tune"], version = "^2.22.0"}
pytorch-lightning = "^2.2.4"
torch = "2.2.2"
geobench = "^1.0.0"
bitsandbytes = "0.41.0"



[tool.poetry.group.style.dependencies]
black = {version = "22.1.0", extras = ["jupyter"]}
flake8 = "6.1.0"
isort = {version = "5.12.0", extras = ["colors"]}
pydocstyle = {version = "6.3.0", extras = ["toml"]}
pyupgrade = "3.15.0"


[tool.poetry.group.datasets.dependencies]
h5py = "^3.10.0"
laspy = "2.5.1"
zipfile-deflate64 = "0.2.0"
rarfile = "4.1"
scikit-image = "0.22.0"
scipy = "1.11.4"


[tool.poetry.group.tests.dependencies]
mypy = "1.6.1"
nbmake = "1.4.6"
pytest = "7.4.3"
pytest-cov = "4.1.0"


[tool.poetry.group.docs.dependencies]
ipywidgets = "8.1.1"
nbsphinx = "0.9.3"
Sphinx = "5.3.0"


[tool.poetry.group.local.dependencies]
jupyter = "^1.0.0"
plotly = "^5.18.0"
ipywidgets = "^8.1.1"

[tool.poetry.group.commit.dependencies]
ruamel-yaml = ">=0.15"
pre-commit = "^3.5.0"
black = "22.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"